+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog of lizard-neerslagradar
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added regions; regions can be added by a script, connected to users,
+  users can login and then zoom to their region, map animation will be
+  confined to their region only.
 
 
 0.1 (2012-11-26)